integrate test pages code  to main dart
